--- boxsplitdb/BoxSplitProject/data/mongo_setup.py
@@ -1,4 +1,0 @@
-import mongoengine
-
-def global_init():
-	mongoengine.register_connection(alias = 'core', name = 'boxsplit')
