+++ python-api/requirements.txt
@@ -1,9 +1,10 @@
 Click==7.0
-fastapi==0.38.1
+fastapi==0.42.0
 h11==0.8.1
-httptools==0.0.13
+mongoengine==0.18.2
 pydantic==0.32.2
-starlette==0.12.8
-uvicorn==0.9.0
-uvloop==0.13.0
-websockets==8.0.2
+pymongo==3.9.0
+six==1.13.0
+starlette==0.12.9
+uvicorn==0.10.8
+websockets==8.1
